requirements.txt
streamlit
groq
python-dotenv
sentence-transformers
scikit-learn
torch
transformers
numpy
google-genai
pillow
bert-score